+++ core/src/com/unciv/models/ruleset/unique/Unique.kt
@@ -23,6 +23,7 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
     /** Does not include conditional params */
     val params = text.getPlaceholderParameters()
     val type = UniqueType.uniqueTypeMap[placeholderText]
+    val deprecatedType: DeprecatedUniqueType? = if (type == null) DeprecatedUniqueType.uniqueTypeMap[placeholderText] else null
 
     val stats: Stats by lazy {
         val firstStatParam = params.firstOrNull { Stats.isStats(it) }
@@ -35,8 +36,8 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
     val isTimedTriggerable = hasModifier(UniqueType.ConditionalTimedUnique)
 
     val isTriggerable = type != null && (
-        type.targetTypes.contains(UniqueTarget.Triggerable)
-            || type.targetTypes.contains(UniqueTarget.UnitTriggerable)
+        type.targetTypes.any { it.canAcceptUniqueTarget(UniqueTarget.Triggerable) }
+            || type.targetTypes.any { it.canAcceptUniqueTarget(UniqueTarget.UnitTriggerable) }
             || isTimedTriggerable
         )
 
@@ -46,7 +47,9 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
     val isLocalEffect = params.contains("in this city") || hasModifier(UniqueType.ConditionalInThisCity)
     val isOtherModifierType = type?.targetTypes?.any { it.modifierType == UniqueTarget.ModifierType.Other } == true
 
-    @Readonly fun hasFlag(flag: UniqueFlag) = type != null && type.flags.contains(flag)
+    @Readonly fun hasFlag(flag: UniqueFlag) =
+        (type != null && type.flags.contains(flag)) ||
+        (deprecatedType != null && deprecatedType.flags.contains(flag))
     @Readonly fun isHiddenToUsers() = hasFlag(UniqueFlag.HiddenToUsers) || hasModifier(UniqueType.ModifierHiddenFromUsers)
 
 
@@ -149,7 +152,8 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
         }
     }
 
-    @Readonly fun getDeprecationAnnotation(): Deprecated? = type?.getDeprecationAnnotation()
+    @Readonly fun getDeprecationAnnotation(): Deprecated? =
+        type?.getDeprecationAnnotation() ?: deprecatedType?.getDeprecationAnnotation()
 
     @Readonly
     fun getSourceNameForUser(): String {
@@ -181,7 +185,7 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
     fun getReplacementText(ruleset: Ruleset): String {
         val deprecationAnnotation = getDeprecationAnnotation() ?: return ""
         val replacementUniqueText = deprecationAnnotation.replaceWith.expression
-        val deprecatedUniquePlaceholders = type!!.text.getPlaceholderParameters()
+        val deprecatedUniquePlaceholders = (type?.text ?: deprecatedType!!.text).getPlaceholderParameters()
         val possibleUniques = replacementUniqueText.split(Constants.uniqueOrDelimiter)
 
         // Here, for once, we DO want the conditional placeholder parameters together with the regular ones,
@@ -207,12 +211,13 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
                 if (parameterNumberInDeprecatedUnique !in params.indices) continue
                 timesParameterWasSeen.add(parameterUnsigned, 1)
                 
-                val positionInDeprecatedUnique =  type.text.indexOf("[$parameterUnsigned]")
+                val positionInDeprecatedUnique = (type?.text ?: deprecatedType!!.text).indexOf("[$parameterUnsigned]")
                 var replacementText = params[parameterNumberInDeprecatedUnique]
-                if (UniqueParameterType.Number in type.parameterTypeMap[parameterNumberInDeprecatedUnique]) {
+                if (type == null || UniqueParameterType.Number in type.parameterTypeMap[parameterNumberInDeprecatedUnique]) {
                     // The following looks for a sign just before [amount] and detects replacing "-[-33]" with "[+33]" and similar situations
-                    val deprecatedHadPlusSign = positionInDeprecatedUnique > 0 && type.text[positionInDeprecatedUnique - 1] == '+'
-                    val deprecatedHadMinusSign = positionInDeprecatedUnique > 0 && type.text[positionInDeprecatedUnique - 1] == '-'
+                    val deprecatedSourceText = type?.text ?: deprecatedType!!.text
+                    val deprecatedHadPlusSign = positionInDeprecatedUnique > 0 && deprecatedSourceText[positionInDeprecatedUnique - 1] == '+'
+                    val deprecatedHadMinusSign = positionInDeprecatedUnique > 0 && deprecatedSourceText[positionInDeprecatedUnique - 1] == '-'
                     val deprecatedHadSign = deprecatedHadPlusSign || deprecatedHadMinusSign
                     val positionInNewUnique = possibleUnique.indexOf("[$parameter]")
                     val newHasMinusSign = positionInNewUnique > 0 && possibleUnique[positionInNewUnique - 1] == '-'