+++ core/src/com/unciv/models/ruleset/validation/RulesetValidator.kt
@@ -36,6 +36,7 @@ import com.unciv.ui.images.AtlasPreview
 import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.images.Portrait
 import com.unciv.ui.images.PortraitPromotion
+import com.unciv.utils.isRunFromJar
 
 /**
  *  Class mananging ruleset validation.
@@ -253,7 +254,7 @@ open class RulesetValidator protected constructor(
                 .filter { it.type == UniqueType.PillageYieldRandom || it.type == UniqueType.PillageYieldFixed }) {
                 if (!Stats.isStats(unique.params[0])) continue
                 val params = Stats.parse(unique.params[0])
-                if (params.values.any { it < 0 }) lines.add(
+                if (params.min() < 0f) lines.add(
                     "${improvement.name} cannot have a negative value for a pillage yield!",
                     RulesetErrorSeverity.Error, improvement
                 )
@@ -676,7 +677,7 @@ open class RulesetValidator protected constructor(
     private fun checkTilesetSanity(lines: RulesetErrorList) {
         // If running from a jar *and* checking a builtin ruleset, skip this check.
         // - We can't list() the jsons, and the unit test before release is sufficient, the tileset config can't have changed since then.
-        if (ruleset.folderLocation == null && this::class.java.`package`?.specificationVersion != null)
+        if (ruleset.folderLocation == null && isRunFromJar(this))
             return
 
         val tilesetConfigFolder = (ruleset.folderLocation ?: Gdx.files.internal("")).child("jsons/TileSets")